archive/dashapps/requirements.txt
alembic==1.9.0
aniso8601==9.0.1
asttokens==2.2.1
async-timeout==4.0.2
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
bleach==5.0.1
blinker==1.5
cachelib==0.10.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dash==2.7.1
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.5
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mailman==0.3.0
Flask-Migrate==4.0.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.0.2
Flask-Session==0.3.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.38.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
Mako==1.2.4
mariadb==1.1.5.post3
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mkdocs-material-extensions==1.1.1
nest-asyncio==1.5.6
numpy==1.24.1
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.12.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pyzmq==25.0.0
redis==4.4.2
requests==2.28.2
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.45
stack-data==0.6.2
tenacity==8.1.0
tornado==6.2
traitlets==5.8.1
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
WTForms==3.0.1
